Реализация V3

+++ src/PriceCalculator.Api/Requests/V1/GoodProperties.cs
@@ -1,7 +1,7 @@
 ﻿namespace PriceCalculator.Api.Requests.V1;
 
 public record GoodProperties(
-    int Length,
-    int Width,
-    int Height
+    decimal Length,
+    decimal Width,
+    decimal Height
     );
